Fix profile routes & adjust lazy loading
