+++ src/app/components/profile/login/login.component.html
@@ -1,1 +1,75 @@
-<p>login works!</p>
+<form class="login-form" [formGroup]="formGroup" (ngSubmit)="login()">
+  <h1>Login</h1>
+  <span>
+    <h2>
+      Don't have an account?
+      <a
+        [routerLink]="'/' + profileRoutes.signUp.path"
+        queryParamsHandling="preserve"
+      >
+        Sign Up
+      </a>
+    </h2>
+  </span>
+
+  <div class="login-fields">
+    <mat-form-field>
+      <mat-label>Username</mat-label>
+      <input
+        matInput
+        [formControlName]="formControls.username"
+        type="email"
+        inputmode="email"
+        autocapitalize="none"
+        autocomplete="off"
+        spellcheck="false"
+      />
+      @if (formGroup.controls.username.invalid) {
+      <mat-error>{{ errors.required(formControls.username) }}</mat-error>
+      }
+    </mat-form-field>
+    <mat-form-field>
+      <mat-label>Password</mat-label>
+      <input
+        matInput
+        [formControlName]="formControls.password"
+        [type]="showPassword() ? 'text' : 'password'"
+        autocapitalize="none"
+        autocomplete="off"
+        spellcheck="false"
+      />
+      <button
+        mat-icon-button
+        matSuffix
+        (click)="togglePasswordVisibility()"
+        [attr.aria-label]="showPassword() ? 'Hide password' : 'Show password'"
+      >
+        <mat-icon>{{
+          showPassword() ? "visibility_off" : "visibility"
+        }}</mat-icon>
+      </button>
+      @if (formGroup.controls.password.invalid) {
+      <mat-error>{{ errors.required(formControls.password) }}</mat-error>
+      }
+    </mat-form-field>
+  </div>
+
+  <h3>
+    <a [routerLink]="'/' + profileRoutes.forgotPassword.path"
+      >Forgot Password?</a
+    >
+  </h3>
+  <span class="submit-row">
+    <button
+      mat-raised-button
+      color="accent"
+      type="submit"
+      [disabled]="!formGroup.valid || isLoading()"
+    >
+      Login
+    </button>
+    @if (isLoading()) {
+    <mat-spinner diameter="40" class="progress-spinner" />
+    }
+  </span>
+</form>
