+++ src/app/components/profile/login/login.component.scss
@@ -1,0 +1,23 @@
+.login-form {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin: 16px;
+}
+
+.login-fields {
+  display: flex;
+  flex-direction: column;
+}
+
+.submit-row {
+  display: flex;
+  align-items: center;
+  gap: 16px;
+
+  .progress-spinner {
+    // Place next to the submit button, but keep the button centered
+    position: absolute;
+    margin-left: calc(73.4px + 16px); // width of button + gap
+  }
+}
